[UPDATE] refactor of files and architectur of project
