--- enums/http_enum.py
@@ -1,11 +1,0 @@
-class HttpStatus:
-    OK = 200
-    CREATED = 201
-    NO_CONTENT = 204
-    BAD_REQUEST = 400
-    UNAUTHORIZED = 401
-    FORBIDDEN = 403
-    NOT_FOUND = 404
-    METHOD_NOT_ALLOWED = 405
-    CONFLICT = 409
-    INTERNAL_SERVER_ERROR = 500
