+++ models/approach.py
@@ -1,0 +1,3 @@
+class Approach:
+    def run(self, q: str, use_summaries: bool) -> any:
+        raise NotImplementedError
