+++ models/keyphrase_model.py
@@ -1,0 +1,9 @@
+from sqlalchemy import Column, Integer, Text
+from db_config import Base
+
+# EyeColorModel implementation
+class KeyphraseModel(Base):
+    __tablename__ = "keyphrase"
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, nullable=False)
+    keyphrase = Column(Text, default="", nullable=False)
