--- railway.json
@@ -1,9 +1,0 @@
-{
-    "$schema": "https://railway.app/railway.schema.json",
-    "build": {
-      "builder": "NIXPACKS"
-    },
-    "deploy": {
-      "startCommand": "hypercorn app:app --bind \"[::]:$PORT\""
-    }
-  }
