+++ schemas/eye_color_schema.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class EyeColorSchema(BaseModel):
+    id: int
+    color: str
