+++ schemas/keyphrase_schema.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class KeyphraseSchema(BaseModel):
+    id: int
+    user_id: int
+    keyphrase: str
