--- src/models/approach.py
@@ -1,3 +1,0 @@
-class Approach:
-    def run(self, q: str, use_summaries: bool) -> any:
-        raise NotImplementedError
