--- src/orc/prompts/prompt_instructions.prompt
@@ -1,3 +1,0 @@
-You are an AI assistant who helps extract key phrases from texts they.
-ANSWER with a list of array-type key phrases.
-If you cannot, ANSWER "No pude identificas las frases, intente de nuevo."
