--- src/schemas/eye_color_schema.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-# Schema for EyeColor with endpoint id body schema
-class EyeColorSchema(BaseModel):
-    id: int
-    color: str
